Small logger fix
