--- python/.idea/codeStyles/Project.xml
@@ -1,5 +1,0 @@
-<component name="ProjectCodeStyleConfiguration">
-  <code_scheme name="Project" version="173">
-    <option name="RIGHT_MARGIN" value="140" />
-  </code_scheme>
-</component>
